graphrag/graphrag/requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
sentence-transformers>=2.2.2
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
typing-extensions>=4.5.0
asyncio
pypdf
langchain_community
faiss-cpu
spacy
rank_bm25
scikit-learn
fastapi
uvicorn
python-multipart
gremlinpython